added fields in all classes

+++ model/src/main/java/org/neustupov/delivery/app/model/entity/base/AbstractEntity.java
@@ -3,13 +3,13 @@ package org.neustupov.delivery.app.model.entity.base;
 import java.time.LocalDateTime;
 import lombok.Data;
 import lombok.EqualsAndHashCode;
-import org.neustupov.delivery.app.model.entity.accounting.Consumer;
+import org.neustupov.delivery.app.model.entity.consumer.Consumer;
 
 @Data
 @EqualsAndHashCode
 public abstract class AbstractEntity {
 
-  private int id;
+  private Long id;
   private LocalDateTime createdAt;
   private LocalDateTime modifiedAt;
   private Consumer createdBy;
